requirements.txt
bcrypt
djangorestframework
djangorestframework-jwt
django-cors-middleware
django==1.11.4